+++ public/customizer/all.php
@@ -1,0 +1,8 @@
+<?php
+/*
+  Controller for handling Theme Customizer
+*/
+require_once 'customizer.php';
+require_once 'customizer-default.php';
+
+new H_Customizer_Default();