+++ public/h-twig.php
@@ -1,12 +1,14 @@
 <?php
-new H_Twig();
+/*
+  Add extra functions to Twig module
+*/
 class H_Twig {
 
   function __construct() {
     add_filter('get_twig', array($this, 'add_to_twig') );
 
     // enable password-protected post
-    add_filter('timber/post/content/show_password_form_for_protected', function($maybe_show) {
+    add_filter('timber/post/content/show_password_form_for_protected', function( $maybe_show ) {
       return true;
     });
   }
@@ -16,16 +18,11 @@ class H_Twig {
 
     @filter get_twig
   */
-  function add_to_twig($twig) {
+  function add_to_twig( $twig ) {
     $twig->addExtension(new Twig_Extension_StringLoader() );
 
     // Parse Markdown
-    $twig->addFilter('markdown', new Twig_Filter_Function(function($text) {
-      $pd = new Parsedown();
-      $text_compiled = $pd->text($text);
-
-      return do_shortcode($text_compiled);
-    }) );
+    $twig->addFilter('markdown', new Twig_Filter_Function( array( $this, 'filter_markdown') ) );
 
     // only if DEBUG is true
     if (defined('WP_DEBUG') && WP_DEBUG === true) {
@@ -41,6 +38,18 @@ class H_Twig {
     }
 
     return $twig;
+  }
+
+
+  /*
+    Convert Markdown syntax
+    {{ post.content | markdown }}
+  */
+  function filter_markdown( $text ) {
+    $pd = new Parsedown();
+    $text_compiled = $pd->text($text);
+
+    return do_shortcode( $text_compiled );
   }
 
 }